wewlad, worked some more on it.
Added the player and the abillity to move him!

Added status messages! Right now the only used one is when you run into a space that is not empty, it says "You can't move there!"

Tomorrow I might try to implement side scrolling. I have a plan for how to do it.