Changed the status infos at the bottom of the screen to run with the improved linked list data structure. Also fixed a bug where I forgot to update the last enemy in the linked list
The reason that this is better is before I was looping through all of a vector, then deleting an element (usually the first). This would cause all the future indices of the vector to be shifted up one which required copying, etc.

Since I am looping through the vector in order ANYWAY then with a linked list deleting an index doesn't require all the copying and shifting, just rearrangement of the pointers to make them point to the right value.

Linked Lists are better when you don't need the index of the array (usually because you're looping through them all anyway), and you need to add and delete elements after creation. It avoids all the shifting and copying.